+++ ledger/api/character/__init__.py
@@ -1,5 +1,6 @@
 from .journal import LedgerJournalApiEndpoints
 from .ledger import LedgerApiEndpoints
+from .planetary import LedgerPlanetaryApiEndpoints
 from .template import LedgerTemplateApiEndpoints
 
 
@@ -7,3 +8,4 @@ def setup(api):
     LedgerApiEndpoints(api)
     LedgerTemplateApiEndpoints(api)
     LedgerJournalApiEndpoints(api)
+    LedgerPlanetaryApiEndpoints(api)