+++ ledger/urls.py
@@ -5,6 +5,7 @@ from django.urls import path, re_path
 from ledger.api import api
 from ledger.views.character.char_audit import add_char, fetch_memberaudit
 from ledger.views.character.character_ledger import character_admin, character_ledger
+from ledger.views.character.planetary import planetary_index
 from ledger.views.corporation.corp_audit import add_corp
 from ledger.views.corporation.corp_events import (
     create_event,
@@ -56,6 +57,8 @@ urlpatterns = [
     path("events/<int:event_id>/edit/", edit_event, name="edit_event"),
     path("events/<int:event_id>/delete/", delete_event, name="delete_event"),
     path("events/ajax/load_events", load_events, name="load_events"),
+    # -- -- Planetary
+    path("planetary/", planetary_index, name="planetary_index"),
     # -- API System
     re_path(r"^api/", api.urls),
 ]